+++ src/main/java/com/google/sps/servlets/GoServlet.java
@@ -1,0 +1,5 @@
+package com.google.sps.servlets;
+
+public class GoServlet {
+  
+}
