+++ src/main/java/com/google/sps/servlets/TasksSummaryServlet.java
@@ -33,13 +33,17 @@ import java.time.Period;
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Date;
 import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+/** Serves selected information from the User's Tasks Account. */
+@WebServlet("/tasks")
 public class TasksSummaryServlet extends AuthenticatedHttpServlet {
   private final TasksClientFactory tasksClientFactory;
 
@@ -61,7 +65,7 @@ public class TasksSummaryServlet extends AuthenticatedHttpServlet {
   }
 
   /**
-   * Returns a summary of tasks from a user's Tasks account
+   * Returns Tasks from the user's Tasks account
    *
    * @param request Http request from client. Should contain idToken and accessToken
    * @param response 403 if user is not authenticated, list of Tasks otherwise
@@ -77,30 +81,38 @@ public class TasksSummaryServlet extends AuthenticatedHttpServlet {
 
     // Get tasks from Google Tasks
     TasksClient tasksClient = tasksClientFactory.getTasksClient(googleCredential);
-    List<TaskList> taskLists = tasksClient.listTaskLists();
-    List<Task> tasks = tasksClient.listAllTasksInAllTaskLists();
+    List<TaskList> allTaskLists = tasksClient.listTaskLists();
 
     // Initialize Tasks Response
-    List<String> taskListTitles = getTaskListTitles(taskLists);
+    List<Task> tasks;
+
+    String taskLists = request.getParameter("taskLists");
+    if (taskLists == null) {
+      tasks = getAllTasksFromAllTaskLists(tasksClient);
+    } else {
+      List<String> selectedTaskListIds = Arrays.asList(taskLists.split(","));
+      tasks = getAllTasksFromSpecificTaskLists(tasksClient, selectedTaskListIds);
+    }
+
+    Map<String, String> taskListIdsToTitles = getTaskListIdsToTitles(allTaskLists);
     long tasksToCompleteCount = countTasksToComplete(tasks);
     long tasksDueTodayCount = countTasksDueToday(tasks);
     long tasksCompletedTodayCount = countTasksCompletedToday(tasks);
     long tasksOverdueCount = countTasksOverdue(tasks);
     TasksResponse tasksResponse =
         TasksResponse.builder()
-            .taskListTitles(taskListTitles)
+            .taskListIdsToTitles(taskListIdsToTitles)
             .tasksToCompleteCount(tasksToCompleteCount)
             .tasksDueTodayCount(tasksDueTodayCount)
             .tasksCompletedTodayCount(tasksCompletedTodayCount)
             .tasksOverdueCount(tasksOverdueCount)
             .build();
 
-    // Convert tasks to JSON and print to response
     JsonUtility.sendJson(response, tasksResponse);
   }
 
-  private List<String> getTaskListTitles(List<TaskList> taskLists) throws IOException {
-    return taskLists.stream().map(taskList -> taskList.getTitle()).collect(Collectors.toList());
+  private Map<String, String> getTaskListIdsToTitles(List<TaskList> taskLists) throws IOException {
+    return taskLists.stream().collect(Collectors.toMap(TaskList::getId, TaskList::getTitle));
   }
 
   private long countTasksToComplete(List<Task> tasks) {
@@ -145,5 +157,41 @@ public class TasksSummaryServlet extends AuthenticatedHttpServlet {
               return dueDate.isBefore(endOfDay);
             })
         .count();
+  }
+
+  /**
+   * Get the all tasks in all the user's task lists
+   *
+   * @param tasksClient Either a mock TaskClient or a taskClient with a valid credential
+   * @return List of tasks from all task lists in user's account
+   * @throws IOException if an issue occurs with the tasksService
+   */
+  private List<Task> getAllTasksFromAllTaskLists(TasksClient tasksClient) throws IOException {
+    List<TaskList> taskLists = tasksClient.listTaskLists();
+    List<Task> tasks = new ArrayList<>();
+    for (TaskList taskList : taskLists) {
+      tasks.addAll(tasksClient.listTasks(taskList));
+    }
+    return tasks;
+  }
+
+  /**
+   * Get the tasks in the user's task lists with the given task list IDs
+   *
+   * @param tasksClient Either a mock TaskClient or a taskClient with a valid credential
+   * @param taskListTitles List of task list IDs which tasks should be obtained from
+   * @return List of tasks from specified task lists in user's account
+   * @throws IOException if an issue occurs with the tasksService
+   */
+  private List<Task> getAllTasksFromSpecificTaskLists(
+      TasksClient tasksClient, List<String> taskListIds) throws IOException {
+    List<TaskList> taskLists = tasksClient.listTaskLists();
+    List<Task> tasks = new ArrayList<>();
+    for (TaskList taskList : taskLists) {
+      if (taskListIds.contains(taskList.getId())) {
+        tasks.addAll(tasksClient.listTasks(taskList));
+      }
+    }
+    return tasks;
   }
 }
