+++ src/test/java/TasksSummaryServletTest.java
@@ -16,6 +16,7 @@ import com.google.api.services.tasks.model.Task;
 import com.google.api.services.tasks.model.TaskList;
 import com.google.appengine.repackaged.com.google.gson.Gson;
 import com.google.common.collect.ImmutableList;
+import com.google.common.collect.ImmutableMap;
 import com.google.sps.model.AuthenticationVerifier;
 import com.google.sps.model.GmailResponse;
 import com.google.sps.model.TasksClient;
@@ -30,6 +31,7 @@ import java.time.ZoneId;
 import java.time.format.DateTimeFormatter;
 import java.time.temporal.ChronoUnit;
 import java.util.List;
+import java.util.Map;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import org.junit.Assert;
@@ -61,22 +63,25 @@ public final class TasksSummaryServletTest extends AuthenticatedServletTestBase 
   private static final String TASK_TITLE_THREE = "task three";
   private static final String TASK_TITLE_FOUR = "task four";
   private static final String TASK_TITLE_FIVE = "task five";
-  private static final String TASK_LIST_TITLE_ONE = "task list one";
-  private static final String TASK_LIST_TITLE_TWO = "task list two";
+  private static final String TASK_LIST_TITLE_ONE = "task list title one";
+  private static final String TASK_LIST_TITLE_TWO = "task list title two";
+  private static final String TASK_LIST_ID_ONE = "task list id one";
+  private static final String TASK_LIST_ID_TWO = "task list id two";
 
   private static final TaskList TASK_LIST_ONE =
-      new TaskList().setTitle(TASK_LIST_TITLE_ONE).setId("taskListOne");
+      new TaskList().setTitle(TASK_LIST_TITLE_ONE).setId(TASK_LIST_ID_ONE);
   private static final TaskList TASK_LIST_TWO =
-      new TaskList().setTitle(TASK_LIST_TITLE_TWO).setId("taskListTwo");
+      new TaskList().setTitle(TASK_LIST_TITLE_TWO).setId(TASK_LIST_ID_TWO);
 
   private static final List<TaskList> NO_TASK_LISTS = ImmutableList.of();
   private static final List<TaskList> ONE_TASK_LIST = ImmutableList.of(TASK_LIST_ONE);
   private static final List<TaskList> TWO_TASK_LISTS =
       ImmutableList.of(TASK_LIST_ONE, TASK_LIST_TWO);
-  private static final List<String> NO_TASK_LISTS_TITLES = ImmutableList.of();
-  private static final List<String> ONE_TASK_LIST_TITLES = ImmutableList.of(TASK_LIST_TITLE_ONE);
-  private static final List<String> TWO_TASK_LISTS_TITLES =
-      ImmutableList.of(TASK_LIST_TITLE_ONE, TASK_LIST_TITLE_TWO);
+  private static final Map<String, String> NO_TASK_LIST_IDS_TO_TITLES = ImmutableMap.of();
+  private static final Map<String, String> ONE_TASK_LIST_IDS_TO_TITLES =
+      ImmutableMap.of(TASK_LIST_ID_ONE, TASK_LIST_TITLE_ONE);
+  private static final Map<String, String> TWO_TASK_LIST_IDS_TO_TITLES =
+      ImmutableMap.of(TASK_LIST_ID_ONE, TASK_LIST_TITLE_ONE, TASK_LIST_ID_TWO, TASK_LIST_TITLE_TWO);
 
   private static final DateTimeFormatter FORMATTER =
       DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ssXXX").withZone(ZoneId.systemDefault());
@@ -112,7 +117,7 @@ public final class TasksSummaryServletTest extends AuthenticatedServletTestBase 
 
   private static final TasksResponse NO_TASK_LISTS_RESPONSE =
       TasksResponse.builder()
-          .taskListTitles(NO_TASK_LISTS_TITLES)
+          .taskListIdsToTitles(NO_TASK_LIST_IDS_TO_TITLES)
           .tasksToCompleteCount(0)
           .tasksDueTodayCount(0)
           .tasksCompletedTodayCount(0)
@@ -120,7 +125,7 @@ public final class TasksSummaryServletTest extends AuthenticatedServletTestBase 
           .build();
   private static final TasksResponse NO_TASKS_RESPONSE =
       TasksResponse.builder()
-          .taskListTitles(ONE_TASK_LIST_TITLES)
+          .taskListIdsToTitles(ONE_TASK_LIST_IDS_TO_TITLES)
           .tasksToCompleteCount(0)
           .tasksDueTodayCount(0)
           .tasksCompletedTodayCount(0)
@@ -128,7 +133,7 @@ public final class TasksSummaryServletTest extends AuthenticatedServletTestBase 
           .build();
   private static final TasksResponse TASKS_DUE_YESTERDAY_RESPONSE =
       TasksResponse.builder()
-          .taskListTitles(ONE_TASK_LIST_TITLES)
+          .taskListIdsToTitles(ONE_TASK_LIST_IDS_TO_TITLES)
           .tasksToCompleteCount(1)
           .tasksDueTodayCount(0)
           .tasksCompletedTodayCount(0)
@@ -136,7 +141,7 @@ public final class TasksSummaryServletTest extends AuthenticatedServletTestBase 
           .build();
   private static final TasksResponse TASKS_DUE_TODAY_RESPONSE =
       TasksResponse.builder()
-          .taskListTitles(ONE_TASK_LIST_TITLES)
+          .taskListIdsToTitles(ONE_TASK_LIST_IDS_TO_TITLES)
           .tasksToCompleteCount(1)
           .tasksDueTodayCount(1)
           .tasksCompletedTodayCount(0)
@@ -144,7 +149,7 @@ public final class TasksSummaryServletTest extends AuthenticatedServletTestBase 
           .build();
   private static final TasksResponse TASKS_DUE_TOMORROW_RESPONSE =
       TasksResponse.builder()
-          .taskListTitles(ONE_TASK_LIST_TITLES)
+          .taskListIdsToTitles(ONE_TASK_LIST_IDS_TO_TITLES)
           .tasksToCompleteCount(1)
           .tasksDueTodayCount(0)
           .tasksCompletedTodayCount(0)
@@ -152,7 +157,7 @@ public final class TasksSummaryServletTest extends AuthenticatedServletTestBase 
           .build();
   private static final TasksResponse TASKS_COMPLETED_YESTERDAY_RESPONSE =
       TasksResponse.builder()
-          .taskListTitles(ONE_TASK_LIST_TITLES)
+          .taskListIdsToTitles(ONE_TASK_LIST_IDS_TO_TITLES)
           .tasksToCompleteCount(0)
           .tasksDueTodayCount(0)
           .tasksCompletedTodayCount(0)
@@ -160,7 +165,7 @@ public final class TasksSummaryServletTest extends AuthenticatedServletTestBase 
           .build();
   private static final TasksResponse TASKS_COMPLETED_TODAY_RESPONSE =
       TasksResponse.builder()
-          .taskListTitles(ONE_TASK_LIST_TITLES)
+          .taskListIdsToTitles(ONE_TASK_LIST_IDS_TO_TITLES)
           .tasksToCompleteCount(0)
           .tasksDueTodayCount(0)
           .tasksCompletedTodayCount(1)
@@ -168,7 +173,7 @@ public final class TasksSummaryServletTest extends AuthenticatedServletTestBase 
           .build();
   private static final TasksResponse ALL_TASKS_RESPONSE =
       TasksResponse.builder()
-          .taskListTitles(ONE_TASK_LIST_TITLES)
+          .taskListIdsToTitles(ONE_TASK_LIST_IDS_TO_TITLES)
           .tasksToCompleteCount(3)
           .tasksDueTodayCount(1)
           .tasksCompletedTodayCount(1)
@@ -176,7 +181,7 @@ public final class TasksSummaryServletTest extends AuthenticatedServletTestBase 
           .build();
   private static final TasksResponse MULTIPLE_TASK_LISTS_RESPONSE =
       TasksResponse.builder()
-          .taskListTitles(TWO_TASK_LISTS_TITLES)
+          .taskListIdsToTitles(TWO_TASK_LIST_IDS_TO_TITLES)
           .tasksToCompleteCount(6)
           .tasksDueTodayCount(2)
           .tasksCompletedTodayCount(2)
@@ -206,8 +211,6 @@ public final class TasksSummaryServletTest extends AuthenticatedServletTestBase 
         Mockito.mock(
             HttpServletResponse.class,
             AdditionalAnswers.delegatesTo(new HttpServletResponseFake(stringWriter)));
-
-    Mockito.when(request.getCookies()).thenReturn(validCookies);
   }
 
   @Test
